+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Next]
+**Fixed**
+- BigDecimal tests are now properly skipped when BigDecimal is not available, instead of always.
+
 ## [v0.2.2] — 2024-10-07
 **Added**
 - Add `#abs` (aliased as `#magnitude`) and `#abs2`.